+++ changes.md
@@ -1,9 +1,10 @@
 # JSROOT changelog
 
-## Changes in master (3.5 dev)
+## Changes in master
 1. Fix error in vertical text alignment
 2. Many improvements in TPaletteAxis drawing - draw label, avoid too large ticks.
 3. Fix error with col drawing - bin with maximum value got wrong color  
+4. Test for existing jquery and jquery-ui libraries, reuse when exists 
 
 
 ## Changes in 3.4